requirements.txt
pyopenssl
asgiref==3.6.0
astroid==2.13.2
asyncio==3.4.3
attrs==21.4.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
cleo==0.8.1
click==8.1.3
clikit==0.6.2
coreapi==2.3.3
coreschema==0.0.4
coverage==7.0.3
crashtest==0.3.1
cryptography==37.0.2
cssutils==2.6.0
dateutils==0.6.12
dill==0.3.6
distlib==0.3.4
Django==4.1.5
django-configurations==2.3.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.1.5
django-filter==22.1
django-nose==1.4.7
django-rest-framework-braces==0.3.4
django_debug_toolbar==3.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-generators==0.5.0
drf-spectacular==0.22.1
drf-spectacular-sidecar==2022.7.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==16.6.0
filelock==3.9.0
filetype==1.1.0
html5lib==1.1
idna==3.3
importlib-metadata==6.0.0
importlib-resources==5.10.2
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.13.1
lazy-object-proxy==1.7.1
lockfile==0.12.2
MarkupSafe==2.1.1
mccabe==0.7.0
model-bakery==1.9.0
more-itertools==9.0.0
msgpack==1.0.4
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
nose==1.3.7
numpy==1.24.1
packaging==20.9
pandas==1.5.2
pastel==0.2.1
pathspec==0.9.0
pexpect==4.8.0
Pillow==9.4.0
pkginfo==1.8.2
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
plazy==0.1.5
pluggy==1.0.0
protobuf==3.20.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.11.0
pyaml==21.10.1
pycparser==2.21
pydantic==1.9.2
PyJWT==2.4.0
pylev==1.4.0
pylint==2.15.10
pyotp==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-http-client==3.3.7
python-magic==0.4.27
python-on-whales==0.55.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.2
sendgrid==6.9.7
shellingham==1.4.0
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tomli==2.0.1
tomlkit==0.11.0
tqdm==4.64.0
typer==0.6.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
webencodings==0.5.1
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.13.0
zipp==3.11.0
ruamel.yaml
phonenumbers==8.13.15
openpyxl==3.0.10
testcontainers==3.7.1
sqlalchemy>=1.3,<2.0
secrets==1.0.2
celery
redis
